Updated hyperverse hook
