+++ packages/hyperverse-evm-tribes/package.json
@@ -31,7 +31,7 @@
     "@openzeppelin/contracts": "^4.4.2",
     "dotenv": "^16.0.0",
     "ethers": "^5.5.3",
-    "hardhat": "^2.9.0",
+    "hardhat": "^2.9.1",
     "react-query": "^3.34.7",
     "react-use": "^17.3.2",
     "skynet-js": "^4.0.23-beta"
@@ -42,7 +42,6 @@
     "@nomiclabs/hardhat-waffle": "^2.0.2",
     "chai": "^4.3.4",
     "ethereum-waffle": "^3.4.0",
-    "hardhat": "^2.8.3",
     "parcel": "^2.2.1",
     "rimraf": "^3.0.2"
   }
