+++ packages/hyperverse/source/index.ts
@@ -1,8 +1,8 @@
 export { default as initialize } from './initialize';
-export { Provider, useHyperverse } from './Provider';
+export { useHyperverse } from './useHyperverse';
 export { useDeviceDetect } from './components';
-import { Provider } from './Provider'
-export { Provider as Hyperverse };
+import { Provider } from './Provider';
+export { Provider, Provider as Hyperverse };
 
 export * from './constants/blockchains';
 export * from './constants/networks';
